// InventarioWebApp/src/app/core/models/CodigoMedico/solicitudCodigo.ts
export class SolicitudCodigoMedicoModel {

    rut: string | undefined;
    nombres: string | undefined;
    nombre2: string | undefined;
    apellpat: string | undefined;
    apellmat: string | undefined;
    categoriaId: number | undefined;
    servicioId: number | undefined;
    fechaInicio: Date | undefined;
    fechaFin: Date | undefined;
    correo: string | undefined;
    certificadoTitulo: string | undefined;
    superIntendencia: string | undefined;
    eunacom: string | undefined;
    certificadoReconocimiento: string | undefined;
  }